refactor: extract CopyButton and add to contract ID display
